requirements.txt
pyTelegramBotAPI>=4.4.0
python-pam>=2.0.2